azure_functions/video_processing_function/requirements.txt
azure-functions
python-dotenv
azure-cosmos
requests
